+++ Challenger.DbUp/DB/Scripts/02/003_dbo.Ingridient.sql
@@ -3,10 +3,10 @@ CREATE TABLE [dbo].[Ingridients](
 	[DishId] [bigint] NOT NULL,
 	[ProductId] [bigint] NOT NULL,
 	[Size] [int] NOT NULL,
-	[Energy] [int] NOT NULL,
-	[Fats] [int] NOT NULL,
-	[Proteins] [int] NOT NULL,
-	[Carbohydrates] [int] NOT NULL,
+	[Energy] [float] NOT NULL,
+	[Fats] [float] NOT NULL,
+	[Proteins] [float] NOT NULL,
+	[Carbohydrates] [float] NOT NULL,
  CONSTRAINT [PK_dbo.Ingridients] PRIMARY KEY CLUSTERED 
 (
 	[Id] ASC
