+++ Challenger.DbUp/DB/Scripts/02/006_dbo.MealProduct.sql
@@ -3,10 +3,10 @@ CREATE TABLE [dbo].[MealProducts](
 	[ProductId] [bigint] NOT NULL,	
 	[MealRecordId] [bigint] NOT NULL,	
 	[Size] [int] NOT NULL,	
-	[Energy] [int] NOT NULL,
-	[Fats] [int] NOT NULL,
-	[Proteins] [int] NOT NULL,
-	[Carbohydrates] [int] NOT NULL,
+	[Energy] [float] NOT NULL,
+	[Fats] [float] NOT NULL,
+	[Proteins] [float] NOT NULL,
+	[Carbohydrates] [float] NOT NULL,
  CONSTRAINT [PK_dbo.MealProducts] PRIMARY KEY CLUSTERED 
 (
 	[Id] ASC
