+++ Challenger.Domain/DbModels/Dish.cs
@@ -25,12 +25,12 @@ namespace Challenger.Domain.DbModels
 
         public List<Ingridient> Ingridients { get; set; }
 
-        public int Energy { get; set; }
+        public double Energy { get; set; }
 
-        public int Fats { get; set; }
+        public double Fats { get; set; }
 
-        public int Proteins { get; set; }
+        public double Proteins { get; set; }
 
-        public int Carbohydrates { get; set; }
+        public double Carbohydrates { get; set; }
     }
 }
