+++ Challenger.Domain/DbModels/Product.cs
@@ -20,12 +20,12 @@ namespace Challenger.Domain.DbModels
         [MaxLength(1)]
         public string SizeUnit { get; set; }
 
-        public int Energy { get; set; }
+        public double Energy { get; set; }
 
-        public int Fats { get; set; }
+        public double Fats { get; set; }
 
-        public int Proteins { get; set; }
+        public double Proteins { get; set; }
 
-        public int Carbohydrates { get; set; }
+        public double Carbohydrates { get; set; }
     }
 }
