+++ Challenger.Domain/Dtos/MealProductDto.cs
@@ -14,12 +14,12 @@ namespace Challenger.Domain.Dtos
 
         public int Size { get; set; }
 
-        public int Energy { get; set; }
+        public double Energy { get; set; }
 
-        public int Fats { get; set; }
+        public double Fats { get; set; }
 
-        public int Proteins { get; set; }
+        public double Proteins { get; set; }
 
-        public int Carbohydrates { get; set; }
+        public double Carbohydrates { get; set; }
     }
 }
